requirements.txt
pycaret==2.2.2
numpy==1.19.2
streamlit==0.71.0
pandas==1.1.2
Pillow==8.1.0
